requirements.txt
grpcio==1.48.2
grpcio-tools==1.48.2
protobuf>=3.12.0,<4.0dev
opencv-python==4.6.0.66
numpy==1.23.5
cython
